Effect size statistics for anova

Returns the (partial) eta-squared, (partial) omega-squared, epsilon-squared statistic or Cohen's F for all terms in an anovas. anova_stats() returns a tidy summary, including all these statistics and power for each term.

anova_stats(model, digits = 3)

Arguments

model

A fitted anova-model of class aov or anova. Other models are coerced to anova.

digits

Amount of digits for returned values.

Value

A data frame with all statistics is returned (excluding confidence intervals).

References

Levine TR, Hullett CR (2002): Eta Squared, Partial Eta Squared, and Misreporting of Effect Size in Communication Research.

Tippey K, Longnecker MT (2016): An Ad Hoc Method for Computing Pseudo-Effect Size for Mixed Model.

Examples

# load sample data
data(efc)

# fit linear model
fit <- aov(
  c12hour ~ as.factor(e42dep) + as.factor(c172code) + c160age,
  data = efc
)
# \dontrun{
anova_stats(car::Anova(fit, type = 2))
#> term                |     sumsq |    meansq |  df | statistic | p.value | etasq | partial.etasq | omegasq | partial.omegasq | epsilonsq | cohens.f | power
#> ----------------------------------------------------------------------------------------------------------------------------------------------------------
#> as.factor(e42dep)   | 4.265e+05 | 1.422e+05 |   3 |    80.299 |  < .001 | 0.212 |         0.224 |   0.209 |           0.221 |     0.209 |    0.537 | 1.000
#> as.factor(c172code) |  7352.049 |  3676.025 |   2 |     2.076 |   0.126 | 0.004 |         0.005 |   0.002 |           0.003 |     0.002 |    0.071 | 0.429
#> c160age             | 1.052e+05 | 1.052e+05 |   1 |    59.408 |  < .001 | 0.052 |         0.066 |   0.051 |           0.065 |     0.051 |    0.267 | 1.000
#> Residuals           | 1.476e+06 |  1770.307 | 834 |           |         |       |               |         |                 |           |          |      
# }