Chi-Squared test

This function performs a \(chi\)^2 test for contingency tables or tests for given probabilities. The returned effects sizes are Cramer's V for tables with more than two rows and columns, Phi (\(\phi\)) for 2x2 tables, and פ (Fei) for tests against given probabilities (see Ben-Shachar et al. 2023).

chi_squared_test(
  data,
  select = NULL,
  by = NULL,
  probabilities = NULL,
  weights = NULL,
  paired = FALSE,
  ...
)

Arguments

data

A data frame.

select

Name of the dependent variable (as string) to be used for the test. select can also be a character vector, specifying the names of multiple continuous variables. In this case, by is ignored and variables specified in select are used to compute the test. This can be useful if the data is in wide-format and no grouping variable is available.

by

Name of the grouping variable to be used for the test. If by is not a factor, it will be coerced to a factor. For chi_squared_test(), if probabilities is provided, by must be NULL.

probabilities

A numeric vector of probabilities for each cell in the contingency table. The length of the vector must match the number of cells in the table, i.e. the number of unique levels of the variable specified in select. If probabilities is provided, a chi-squared test for given probabilities is conducted. Furthermore, if probabilities is given, by must be NULL. The probabilities must sum to 1.

weights

Name of an (optional) weighting variable to be used for the test.

paired

Logical, if TRUE, a McNemar test is conducted for 2x2 tables. Note that paired only works for 2x2 tables.

...

Additional arguments passed down to chisq.test().

Value

A data frame with test results. The returned effects sizes are Cramer's V for tables with more than two rows and columns, Phi (\(\phi\)) for 2x2 tables, and פ (Fei) for tests against given probabilities.

Details

The function is a wrapper around chisq.test() and fisher.test() (for small expected values) for contingency tables, and chisq.test() for given probabilities. When probabilities are provided, these are rescaled to sum to 1 (i.e. rescale.p = TRUE). When fisher.test() is called, simulated p-values are returned (i.e. simulate.p.value = TRUE, see ?fisher.test). If paired = TRUE and a 2x2 table is provided, a McNemar test (see mcnemar.test()) is conducted.

The weighted version of the chi-squared test is based on the a weighted table, using xtabs() as input for chisq.test().

References

Ben-Shachar, M.S., Patil, I., Thériault, R., Wiernik, B.M., Lüdecke, D. (2023). Phi, Fei, Fo, Fum: Effect Sizes for Categorical Data That Use the Chi‑Squared Statistic. Mathematics, 11, 1982. doi:10.3390/math11091982

Examples

data(efc)
efc$weight <- abs(rnorm(nrow(efc), 1, 0.3))

# Chi-squared test
chi_squared_test(efc, "c161sex", by = "e16sex")
#> 
#> # Chi-squared test for contingency tables
#> 
#>   Data: c161sex by e16sex (n = 900)
#> 
#>   χ² = 2.2327, ϕ = 0.0526, df = 1, p = 0.135
#> 

# weighted Chi-squared test
chi_squared_test(efc, "c161sex", by = "e16sex", weights = "weight")
#> 
#> # Chi-squared test for contingency tables (weighted)
#> 
#>   Data: c161sex by e16sex (n = 926)
#> 
#>   χ² = 2.7855, ϕ = 0.0576, df = 1, p = 0.095
#> 

# Chi-squared test for given probabilities
chi_squared_test(efc, "c161sex", probabilities = c(0.3, 0.7))
#> 
#> # Chi-squared test for given probabilities
#> 
#>   Data: c161sex against probabilities 30% and 70% (n = 901)
#> 
#>   χ² = 16.1624, פ‎ = 0.0877, df = 1, p < .001
#>