Get relative risks estimates from logistic regressions or odds ratio values

odds_to_rr() converts odds ratios from a logistic regression model (including mixed models) into relative risks; or_to_rr() converts a single odds ratio estimate into a relative risk estimate.

odds_to_rr(fit)

or_to_rr(or, p0)

Arguments

fit
A fitted binomial generalized linear (mixed) model with logit-link function (logistic (multilevel) regression model).

or
Numeric, an odds ratio estimate.

p0
Numeric, the risk of having a positive outcome in the control or unexposed group (reference group), i.e. the number of outcome or "successes" in the control divided by the total number of observations in the control group.

Value

A data frame with relative risks and lower/upper confidence interval for the relative risks estimates; for or_to_rr(), the risk ratio estimate.

Details

This function extracts the odds ratios (exponentiated model coefficients) from logistic regressions (fitted with glm or glmer) and their related confidence intervals, and transforms these values into relative risks (and their related confidence intervals).

The formula for transformation is based on Zhang and Yu (1998), Wang (2013) and Grant (2014): RR <- OR / (1 - P0 + (P0 * OR)), where OR is the odds ratio and P0 indicates the proportion of the incidence in the outcome variable for the control group (reference group).

References

Grant RL. 2014. Converting an odds ratio to a range of plausible relative risks for better communication of research findings. BMJ 348:f7450. doi: 10.1136/bmj.f7450

Wang Z. 2013. Converting Odds Ratio to Relative Risk in Cohort Studies with Partial Data Information. J Stat Soft 2013;55. doi: 10.18637/jss.v055.i05

Zhang J, Yu KF. 1998. What's the Relative Risk? A Method of Correcting the Odds Ratio in Cohort Studies of Common Outcomes. JAMA; 280(19): 1690-1. doi: 10.1001/jama.280.19.1690

Examples

library(sjmisc)
library(lme4)

#> Loading required package: Matrix

# create binary response
sleepstudy$Reaction.dicho <- dicho(sleepstudy$Reaction, dich.by = "median")
# fit model
fit <- glmer(Reaction.dicho ~ Days + (Days | Subject),
             data = sleepstudy, family = binomial("logit"))
# convert to relative risks
odds_to_rr(fit)

#>     Parameter Odds Ratio Risk Ratio      CI_low   CI_high
#> 1 (Intercept) 0.02201714 0.04308565 0.004411356 0.3597023
#> 2        Days 2.43719045 1.41812942 1.212151446 1.5885357

data(efc)
# create binary response
y <- ifelse(efc$neg_c_7 < median(na.omit(efc$neg_c_7)), 0, 1)
# create data frame for fitted model
mydf <- data.frame(
  y = as.factor(y),
  sex = to_factor(efc$c161sex),
  dep = to_factor(efc$e42dep),
  barthel = efc$barthtot,
  education = to_factor(efc$c172code)
)
# fit model
fit <- glm(y ~., data = mydf, family = binomial(link = "logit"))
# convert to relative risks
odds_to_rr(fit)

#> Waiting for profiling to be done...

#>     Parameter Odds Ratio Risk Ratio    CI_low   CI_high
#> 1 (Intercept)   2.014220  1.2942650 0.8005943 1.6101037
#> 2        sex2   1.913887  1.3686646 1.1632760 1.5625768
#> 3        dep2   1.624596  1.4468228 0.8511852 2.3180826
#> 4        dep3   3.080617  2.1859172 1.4029261 3.1216957
#> 5        dep4   2.484804  1.9230853 1.0500095 3.0354848
#> 6     barthel   0.970638  0.9865251 0.9822535 0.9906641
#> 7  education2   1.254234  1.1027736 0.9233955 1.2698368
#> 8  education3   1.327901  1.1280723 0.9057139 1.3294549

# replicate OR/RR for coefficient "sex" from above regression
# p0 ~ .44, or ~ 1.914
prop.table(table(mydf$y, mydf$sex))

#>    
#>             1         2
#>   0 0.1324355 0.3153760
#>   1 0.1054994 0.4466891

or_to_rr(1.914, 0.1055 / (.1324 + .1055))

#> [1] 1.361962