Compute model quality

Compute root mean squared error, residual standard error or mean square error of fitted linear (mixed effects) models.

rmse(fit, normalized = FALSE)

rse(fit)

mse(fit)

Arguments

fit
Fitted linear model of class lm, merMod (lme4) or lme (nlme).

normalized
Logical, use TRUE if normalized rmse should be returned.

Note

Root Mean Square Error

The RMSE is the square root of the variance of the residuals and indicates the absolute fit of the model to the data (difference between observed data to model's predicted values). “RMSE can be interpreted as the standard deviation of the unexplained variance, and has the useful property of being in the same units as the response variable. Lower values of RMSE indicate better fit. RMSE is a good measure of how accurately the model predicts the response, and is the most important criterion for fit if the main purpose of the model is prediction.” (Grace-Martin K: Assessing the Fit of Regression Models)

The normalized RMSE is the proportion of the RMSE related to the range of the response variable. Hence, lower values indicate less residual variance.

Residual Standard Error

The residual standard error is the square root of the residual sum of squares divided by the residual degrees of freedom.

Mean Square Error

The mean square error is the mean of the sum of squared residuals, i.e. it measures the average of the squares of the errors. Lower values (closer to zero) indicate better fit.

References

Grace-Martin K: Assessing the Fit of Regression Models

See also

Examples

data(efc)
fit <- lm(barthtot ~ c160age + c12hour, data = efc)
rmse(fit)

#> [1] 25.46792

rse(fit)

#> [1] 25.51134

library(lme4)
fit <- lmer(Reaction ~ Days + (Days | Subject), sleepstudy)
rmse(fit)

#> [1] 23.43806

mse(fit)

#> [1] 549.3425

# normalized RMSE
library(nlme)
fit <- lme(distance ~ age, data = Orthodont)
rmse(fit, normalized = TRUE)

#> [1] 0.07242178