Mann-Whitney-Test

This function performs a Mann-Whitney-Test (or Wilcoxon rank sum test for unpaired samples.

A Mann-Whitney-Test is a non-parametric test for the null hypothesis that two independent samples have identical continuous distributions. It can be used when the two continuous variables are not normally distributed.

mann_whitney_test(
  data,
  select = NULL,
  by = NULL,
  weights = NULL,
  distribution = "asymptotic"
)

Arguments

data

A data frame.

select

Name of the dependent variable (as string) to be used for the test. select can also be a character vector, specifying the names of multiple continuous variables. In this case, by is ignored and variables specified in select are used to compute the test. This can be useful if the data is in wide-format and no grouping variable is available.

by

Name of the grouping variable to be used for the test. If by is not a factor, it will be coerced to a factor. For chi_squared_test(), if probabilities is provided, by must be NULL.

weights

Name of an (optional) weighting variable to be used for the test.

distribution

Indicates how the null distribution of the test statistic should be computed. May be one of "exact", "approximate" or "asymptotic" (default). See coin::wilcox_test() for details.

Value

A data frame with test results. The function returns p and Z-values as well as effect size r and group-rank-means.

Details

This function is based on wilcox.test() and coin::wilcox_test() (the latter to extract effect sizes). The weighted version of the test is based on survey::svyranktest().

Interpretation of the effect size r, as a rule-of-thumb:

small effect >= 0.1

medium effect >= 0.3

large effect >= 0.5

r is calcuated as:

r = |Z| / sqrt(n1 + n2)

Examples

data(efc)
# Mann-Whitney-U-Tests for elder's age by elder's sex.
mann_whitney_test(efc, "e17age", by = "e16sex")
#> # Mann-Whitney test
#> 
#>   Group 1: male   (n = 294, rank mean = 147.50)
#>   Group 2: female (n = 596, rank mean = 298.50)
#> 
#>   r = 0.257, Z = -7.748, p < .001
#> 

# when data is in wide-format, specify all relevant continuous
# variables in `select` and omit `by`
set.seed(123)
wide_data <- data.frame(scale1 = runif(20), scale2 = runif(20))
mann_whitney_test(wide_data, select = c("scale1", "scale2"))
#> # Mann-Whitney test
#> 
#>   Group 1: scale1 (n = 20, rank mean = 10.50)
#>   Group 2: scale2 (n = 20, rank mean = 10.50)
#> 
#>   r = 0.051, Z = -0.325, p = 0.745
#> 

# same as if we had data in long format, with grouping variable
long_data <- data.frame(
  scales = c(wide_data$scale1, wide_data$scale2),
  groups = rep(c("A", "B"), each = 20)
)
mann_whitney_test(long_data, select = "scales", by = "groups")
#> # Mann-Whitney test
#> 
#>   Group 1: A (n = 20, rank mean = 10.50)
#>   Group 2: B (n = 20, rank mean = 10.50)
#> 
#>   r = 0.051, Z = -0.325, p = 0.745
#>